HCK-9687: Adapt FE to work in the browser
