+++ api/fe.js
@@ -1,0 +1,7 @@
+const { generateContainerScript } = require('../forward_engineering/generateContainerScript');
+const { generateScript } = require('../forward_engineering/generateScript');
+
+module.exports = {
+	generateContainerScript,
+	generateScript,
+};
