+++ package.json
@@ -39,6 +39,7 @@
     "description": "Hackolade plugin for MarkLogic",
     "disabled": false,
     "dependencies": {
+        "async": "3.2.6",
         "lodash": "4.17.21",
         "marklogic": "3.4.0",
         "xml-js": "1.6.11"
@@ -52,7 +53,8 @@
     },
     "scripts": {
         "lint": "eslint . --max-warnings=0",
-        "package": "node esbuild.package.js"
+        "package": "node esbuild.package.js",
+        "postinstall": "npx simple-git-hooks"
     },
     "devDependencies": {
         "@hackolade/hck-esbuild-plugins-pack": "0.0.1",
@@ -60,6 +62,7 @@
         "@typescript-eslint/parser": "7.11.0",
         "esbuild": "0.20.2",
         "esbuild-plugin-clean": "1.0.1",
+        "esbuild-plugin-copy": "2.1.1",
         "eslint": "8.57.0",
         "eslint-config-prettier": "9.1.0",
         "eslint-formatter-teamcity": "^1.0.0",
