--- reverse_engineering/appDependencies.js
@@ -1,11 +1,0 @@
-let dependencies = {};
-
-const setDependencies = app => {
-	dependencies.lodash = app.require('lodash');
-	dependencies.async = app.require('async');
-};
-
-module.exports = {
-	setDependencies,
-	dependencies,
-};
